bug: Return auth_time in ID token claims
This claim is REQUIRED only when a max_age request is made, or when
auth_time is requested as an Essential Claim (which isn't supported yet)

The parameter information from the original authorization is unavailable
in the code request, so auth_time is currently always returned.
To get around this we might want to generate the ID tokens in the
authorization request and destroy them after usage, to replace the
current nonces table.